+++ src/main/java/com/imsweb/naaccrxml/SpecificationVersion.java
@@ -8,17 +8,27 @@ import org.apache.commons.lang3.StringUtils;
 public class SpecificationVersion {
 
     /**
-     * Constants for specification version 1.2
+     * Constant for the current specification implemented by this library
+     */
+    public static final String CURRENT_SPECIFICATION = "1.3";
+
+    /**
+     * Constant for specification version 1.3
+     */
+    public static final String SPEC_1_3 = "1.3";
+
+    /**
+     * Constant for specification version 1.2
      */
     public static final String SPEC_1_2 = "1.2";
 
     /**
-     * Constants for specification version 1.1
+     * Constant for specification version 1.1
      */
     public static final String SPEC_1_1 = "1.1";
 
     /**
-     * Constants for specification version 1.0
+     * Constant for specification version 1.0
      */
     public static final String SPEC_1_0 = "1.0";
 
@@ -26,7 +36,7 @@ public class SpecificationVersion {
      * Returns true if the provided spec is supported by this library, false otherwise.
      */
     public static boolean isSpecificationSupported(String spec) {
-        return SPEC_1_0.equals(spec) || SPEC_1_1.equals(spec) || SPEC_1_2.equals(spec);
+        return SPEC_1_0.equals(spec) || SPEC_1_1.equals(spec) || SPEC_1_2.equals(spec) || SPEC_1_3.equals(spec);
     }
 
     /**